Refactor of User controller and service

+++ src/user/user.service.ts
@@ -1,9 +1,10 @@
-import { Injectable } from "@nestjs/common";
-import { Prisma, User } from "@prisma/client";
+import { ForbiddenException, Injectable, NotFoundException } from "@nestjs/common";
+import { Prisma } from "@prisma/client";
 import { PrismaService } from "src/prisma/prisma.service";
-import { UserResponseDTO } from "./DTOs/UserResponse.dto";
 import { AuthUserDTO } from "src/auth/DTOs/authUser.dto";
-import { UserWithIdDTO } from "./DTOs/UserResponseWithId.dto";
+import { ResponseUserDTO } from "./DTOs/responseUser.dto";
+import { UpdateUserDTO } from "./DTOs/updateUser.dto";
+import { CreateUserDTO } from "./DTOs/createUser.dto";
 
 @Injectable()
 export class UserService {
@@ -11,14 +12,19 @@ export class UserService {
 
     async user(
         userWhereUniqueInput: Prisma.UserWhereUniqueInput,
-    ): Promise<UserResponseDTO | null> {
-        return this.prisma.user.findUnique({
+    ): Promise<ResponseUserDTO> {
+        const user = await this.prisma.user.findUnique({
             where: userWhereUniqueInput,
             select: {
+                id: true,
                 email: true,
                 name: true,
             },
         });
+
+        if (!user) throw new NotFoundException("User not found");
+
+        return user;
     }
 
     async users(params: {
@@ -27,7 +33,7 @@ export class UserService {
         cursor?: Prisma.UserWhereUniqueInput;
         where?: Prisma.UserWhereInput;
         orderBy?: Prisma.UserOrderByWithRelationInput;
-    }): Promise<UserResponseDTO[]> {
+    }): Promise<ResponseUserDTO[]> {
         const { skip, take, cursor, where, orderBy } = params;
         return this.prisma.user.findMany({
             skip,
@@ -36,17 +42,18 @@ export class UserService {
             where,
             orderBy,
             select: {
+                id: true,
                 email: true,
                 name: true,
             },
         });
     }
 
-    async getUserByEmail(
-        where: Prisma.UserWhereUniqueInput,
-    ): Promise<AuthUserDTO | null> {
+    async getUserByEmail(email: string): Promise<AuthUserDTO | null> {
         return this.prisma.user.findUnique({
-            where,
+            where: {
+                email
+            },
             select: {
                 id: true,
                 email: true,
@@ -55,9 +62,9 @@ export class UserService {
         });
     }
 
-    async createUser(data: Prisma.UserCreateInput): Promise<UserWithIdDTO> {
+    async createUser(dto: CreateUserDTO): Promise<ResponseUserDTO> {
         return this.prisma.user.create({
-            data,
+            data: dto,
             select: {
                 id: true,
                 email: true,
@@ -66,27 +73,39 @@ export class UserService {
         });
     }
 
-    async updateUser(params: {
-        where: Prisma.UserWhereUniqueInput;
-        data: Prisma.UserUpdateInput;
-    }): Promise<UserResponseDTO> {
-        const { data, where } = params;
+    async updateUser(id: string, userId: string, dto: UpdateUserDTO): Promise<ResponseUserDTO> {
+        if (userId !== id){
+            throw new ForbiddenException("You can't update this user");
+        }
+
+        await this.user({ id });
+
         return this.prisma.user.update({
-            data,
-            where,
+            where: {
+                id
+            },
+            data: dto,
             select: {
+                id: true,
                 email: true,
                 name: true,
             },
         });
     }
 
-    async deleteUser(
-        where: Prisma.UserWhereUniqueInput,
-    ): Promise<UserResponseDTO> {
+    async deleteUser(id: string, userId: string): Promise<ResponseUserDTO> {
+        if (userId !== id){
+            throw new ForbiddenException("You can't delete this user");
+        }
+
+        await this.user({ id });
+
         return this.prisma.user.delete({
-            where,
+            where: {
+                id
+            },
             select: {
+                id: true,
                 email: true,
                 name: true,
             },
